requirements.txt
python-dotenv
openai
pytubefix
argparse